docs: the g2p help message should be honest - add caveat
This version of "readalongs g2p" produces the same .g2p.xml file that
align produces, which is not usable as input to align, it's more just
for testing and seeing what happens at the moment.

A future version of this function will produce a different g2p format
that will be usable as input to align, but we're not there yet.