feat: save g2p in ARPABET attribute and accept pre-g2p'd text
WIP

readalongs g2p now saves the g2p output to attribute ARPABET in the XML
instead of replacing the text.

readalongs align now accepts a g2p'd input file, or a partially g2p'd
one, and only g2p's the rest of the tokens, accepting the given
pronounciation for <w> elements that already have the ARPABET attrib.

testing for this feature is started but not complete yet.