fix: don't save SoundSwallower logs on Windows, it's buggy
On Windows, you cannot delete a file if you still have a filehandle open
on it, and soundswallower does not close its file handle on ss_log
passed via cfg.set_string("-logfn", ss_log). Not the best solution, but
we just don't save those log files when we're on Windows. Use
--debug-aligner on that platform to see them.