style: quench some mypy error, and change how we install mypy
The pre-commit hook for mypy is not installed via requirements.dev.txt and invoked
locally in .pre-commit-config.yaml, which makes it easier to install all the types-*
dependencies we want.

Bonus: running mypy on the command line and in the pre-commit hook now invokes the
same version, and therefore provide consistent output.