ci: trying to fix readthedocs build
readthedocs is currently broken, because it pip installs readalongs from
PyPI, which gets the latest published version, but pip installs the
current requirements.txt file. At the moment, that's in conflict because
we've updated the version of SoundSwallower we're using, so we're
simultaneously saying ss==0.1.5 and ss==0.2.2, which cannot be resolved,
and breaks compiling the documentation.