style: replace commitlint by gitlint
gitlint is pure python, whereas commitlint is JavaScript and required
node. Thus, we completely remove the dependency on Node and instead use
our dev dependency mechanism that we already had in place.

Plus, gitlint prompts you when you commit is not compliant instead of
aborting right away, which I think is nicer.