fix: extract sentences correctly on page changes
Identifying the sentence ID of words past the first paragraph was
incorrect, resulting in wrong sentence layers for the vtt, srt and
TextGrid output formats.

The sentence id needs to be the whole word's ID up to be excluding the
word ID itself. I.e. d0p0s0w0 and d0p1s0w0 are not part of the same
sentence, since they're in the first sentence of different paragraphs.

while I'm in here, refactor a bit and clean up documentation.

And include unit testing.

FIXES #70