fix: more robust sentence extraction for srt, vtt, TextGrid
My first attempt at fixing sentence extraction was still looking at the
word IDs, assuming that p0s1w2 and p0s1w3 where in the same sentence by
p1s1w2 in a different sentence, but our intent is actually that the user
can provide their own IDs in the first place, and we make a commitment
not to parse IDs or assign semantics to them, so that's still wrong.

This new solution uses the XML structure instead: <w> elements are in
the same sentence if and only if they are under the same <s> element.