docs: quote the pip install -e '.[dev]' so it works everywhere
On a Mac with the default zsh shell configuration, Marc got this error:

$ pip install -e .[dev]
zsh: no matches found: .[dev]

Using .[dev] without quoting is indeed relying on the fact that in bash,
and most shells, if the expansion of .[dev], which is one of .d, .e or
.v, matches no files, the expansion just doesn't happen at all. So we do
in fact rely on the fact that such files don't actually exist when we
run `pip install -e .[dev]` with the last argument unquoted.

I guess in zsh, they decided it was a safety issue or I don't know what,
and made it an error instead.

Quoting prevents the expansion in the first place, and it's safe with
every shell.