wip: when a word g2p's to nothing, convert it to non-word [no ci]
This will make cases where stray diacritics create a spurious word
work in Studio-Web instead of being unalignable

But discussing it with Del, it's not what we want to do, instead we want
to clearly highlight errors.

This branch is going to be throw-away work...

The changes break unit testing all over, and for good reasons. Sigh...